Client/Admin/requirements.txt
weaviate-client
streamlit
st-files-connection
unstructured[all-docs]